requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
transformers
datasets
yfinance
pyyaml
tqdm
scipy
statsmodels
streamlit
altair
